requirements.txt
flask==3.1.0
requests==2.32.3
python-dotenv==1.0.1
pytest==8.3.4
pytest-mock==3.14.0
